Avoid the warning about a double input to abs()
The patch keeps the current behavior, but the warning was much
probably actually raising a bug or a small conceptual mistake:
the type of L_xx.size is float (see tab.h), however it is always
filled with int values because it is initialized to 0 and updated
unsing "int abs(int);". Worse, the input of abs() is always a
double! There's obviously some inconsistency there...